+++ CHANGELOG.md
@@ -13,9 +13,10 @@
   `@fullsoak/fullsoak/batteries`
 - using env var `PREACT_DEVTOOLS` to enable preact devtools when not in `DEBUG`
   mode
-- BREAKING: all init functions (e.g. `useFullSoak()`) now returns a `Promise` of
-  whatever it used to return. This is needed so that the "global components dir"
-  setting is correctly applied dynamically as per user setting
+- BREAKING CHANGE: all init functions (e.g. `useFullSoak()`) now return the
+  `Promise`d version of whatever it used to return directly before. This is
+  needed so that `getGlobalComponentsDir()` works properly in various contexts
+  later on
 
 ### Fixed